:white_check_mark:压力测试

+++ ConfigMgr.h
@@ -8,7 +8,7 @@ class ConfigMgr
 {
 public:
     std::string ip;
-    int port;
+    unsigned short port;
     std::string dbname;
     std::string username;
     std::string password;
